Group 5: Extract HLS service, fix VOD URL path, add reachability cache
- Move _build_hls_url and _resolve_hls_url out of timeline.py into
  new backend/app/services/hls.py so both preview.py and timeline.py
  can import them without a circular dependency
- Fix VOD URL path: /vod/ → /api/vod/ (bare /vod/ returns nginx 400)
- Add 30s per-camera reachability cache to _resolve_hls_url so rapid
  scrubbing doesn't fire a HEAD request on every seek
- Update test patches from app.routers.timeline to app.services.hls
- Update CLAUDE.md: mark v2 bugs resolved, document v3 known issues
  and planned features, update test structure
